requirements-django4.txt
Django==4.0.5
moesifdjango==2.0.5
djangorestframework==3.13.1
